+++ Changelog.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Add new mutation of `Array(foo)` -> `[foo]` [#1039](https://github.com/mbj/mutant/pull/1039)
+
 # v0.9.11 2020-08-25
 
 * Remove mutation to equivalent semantics on endless ranges [#1036](https://github.com/mbj/mutant/pull/1036).